fix: clippy --all-targets lints and minix-libc sendto type error
Clean up the clippy warnings that surface with
`--all-targets -- -D warnings` (test targets were previously not
linted),
and fix a minix-target compile error in minix-libc.

- drivers/virtio_net.rs: constant `assert!` -> `const { assert!(..) }`;
  drop the unneeded `&mut` on `fifo_len` in tests
- kernel/elf.rs: replace the explicit `si` loop counter with
  `.enumerate()` in the exec-frame test helper
- servers/pm.rs: drop the no-op `signo as usize` cast in the ksig
  bitmask tests
- servers/vfs/glo.rs: move `mod tests` to the end of the file
- servers/vfs/path.rs, protect.rs: `Fproc::default()` + field
  assignment -> struct literal with `..Default::default()`; drop the
  `| 0o000` / `| 0` no-ops
- minix-libc: `sendto` returns `isize` but returned the `i32` errno
  from `decode_sockaddr_in` (E0308 on the minix target); cast to isize

Verified: clippy --workspace --all-targets -- -D warnings is clean and
minix-libc compiles for x86_64-pc-minix; drivers/kernel/servers tests
pass.

+++ crates/minix-libc/src/lib.rs
@@ -324,7 +324,7 @@ pub unsafe extern "C" fn sendto(
     } else {
         match unsafe { decode_sockaddr_in(dest_addr as *const u8, dest_len) } {
             Ok((ip, port)) => Some(minix_std::net::SocketAddr { ip, port }),
-            Err(e) => return e,
+            Err(e) => return e as isize,
         }
     };
     match unsafe { minix_std::net::sendto(sock, slice, dest) } {
